% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch06-additive-models.R
\name{casl_am_backfit}
\alias{casl_am_backfit}
\title{Fit linear additive model using backfit algorithm.}
\usage{
casl_am_backfit(X, y, maxit = 10L)
}
\arguments{
\item{X}{A numeric data matrix.}

\item{y}{Response vector.}

\item{maxit}{Integer maximum number of iterations.}
}
\value{
A list of smoothing spline objects; one for each column
in X.
}
\description{
Fit linear additive model using backfit algorithm.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
