% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch06-additive-models.R
\name{casl_am_ls}
\alias{casl_am_ls}
\title{Additive model using truncated power basis.}
\usage{
casl_am_ls(X, y, lambda, order = 3L, nknots = 5L)
}
\arguments{
\item{X}{A numeric data matrix.}

\item{y}{Response vector.}

\item{lambda}{Vector of penalty terms, with one value per
column in the original data matrix.}

\item{order}{Order of the polynomial.}

\item{nknots}{Number of knots in the piecewise polynomial.}
}
\value{
A vector of predictions; one for each row in X_new.
}
\description{
Additive model using truncated power basis.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
