% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch06-additive-models.R
\name{casl_am_ls_f}
\alias{casl_am_ls_f}
\title{Compute predicted values for penalized additive models.}
\usage{
casl_am_ls_f(Z, beta_hat, p)
}
\arguments{
\item{Z}{A numeric data matrix.}

\item{beta_hat}{Estimated regression vector.}

\item{p}{Number of columns in original data matrix.}
}
\value{
Predicted values, f_j, from the additive model.
}
\description{
Compute predicted values for penalized additive models.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
