% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch06-additive-models.R
\name{casl_am_ls_f_se}
\alias{casl_am_ls_f_se}
\title{Compute standard errors for penalized additive models.}
\usage{
casl_am_ls_f_se(Z, omega, beta_hat, s_hat, p)
}
\arguments{
\item{Z}{A numeric data matrix.}

\item{omega}{A penalty matrix.}

\item{beta_hat}{Estimated regression vector.}

\item{s_hat}{Estimated noise variance.}

\item{p}{Number of columns in original data matrix.}
}
\value{
A matrix of standard errors for the predicted values,
f_j, from the additive model.
}
\description{
Compute standard errors for penalized additive models.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
