% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch06-additive-models.R
\name{casl_am_ls_omega}
\alias{casl_am_ls_omega}
\title{Compute penalty matrix for penalized additive models.}
\usage{
casl_am_ls_omega(Z, lambda)
}
\arguments{
\item{Z}{A numeric data matrix.}

\item{lambda}{Vector of penalty terms, with one value per
column in the original data matrix.}
}
\value{
The penalty matrix for the additive model.
}
\description{
Compute penalty matrix for penalized additive models.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
