% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch06-additive-models.R
\name{casl_am_predict}
\alias{casl_am_predict}
\title{Predict values from linear additive model.}
\usage{
casl_am_predict(models, X_new)
}
\arguments{
\item{models}{A list of smoothing spline objects; one for each
column in the training data matrix.}

\item{X_new}{Numeric data matrix of prediction locations.}
}
\value{
A list of smoothing spline objects; one for each column
in X.
}
\description{
Predict values from linear additive model.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
