% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch10-implementation.R
\name{casl_blockwise_irwls}
\alias{casl_blockwise_irwls}
\title{Compute GLM regression with data stored on disk.}
\usage{
casl_blockwise_irwls(filename, nmax, family = stats::gaussian(),
  maxit = 25L, tol = 1e-08)
}
\arguments{
\item{filename}{Path to a file containing the dataset.}

\item{nmax}{Maximum number of rows in the chunk.}

\item{family}{Instance of an R `family` object.}

\item{maxit}{Integer maximum number of iterations.}

\item{tol}{Numeric tolerance parameter.}
}
\value{
A list containing the regression vector beta and
number of iterations.
}
\description{
Compute GLM regression with data stored on disk.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
