% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch08-neural-networks.R
\name{casl_cnn_backward_prop}
\alias{casl_cnn_backward_prop}
\title{Apply backward propagation algorithm for a CNN.}
\usage{
casl_cnn_backward_prop(x, y, weights, f_obj, sigma_p)
}
\arguments{
\item{x}{A numeric vector representing one row of the input.}

\item{y}{A numeric vector representing one row of the response.}

\item{weights}{A list created by casl_nn_make_weights.}

\item{f_obj}{Output of the function casl_nn_forward_prop.}

\item{sigma_p}{Derivative of the activation function.}
}
\value{
A list containing the new weighted responses (z) and activations (a).
}
\description{
Apply backward propagation algorithm for a CNN.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
