% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch08-neural-networks.R
\name{casl_cnn_make_weights}
\alias{casl_cnn_make_weights}
\title{Create list of weights and momentum to describe a CNN.}
\usage{
casl_cnn_make_weights(sizes)
}
\arguments{
\item{sizes}{A vector giving the size of each layer, including
the input and output layers.}
}
\value{
A list containing initialized weights, biases, and momentum.
}
\description{
Create list of weights and momentum to describe a CNN.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
