% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch10-implementation.R
\name{casl_fhash}
\alias{casl_fhash}
\title{Map text to numeric hash values.}
\usage{
casl_fhash(text, k = 4L)
}
\arguments{
\item{text}{A character vector of inputs.}

\item{k}{Size of the hash, given as 2^(4*k).}
}
\value{
A data frame giving hashed values.
}
\description{
Map text to numeric hash values.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
