% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch10-implementation.R
\name{casl_fhash_cache}
\alias{casl_fhash_cache}
\title{Construct a data matrix using hash function.}
\usage{
casl_fhash_cache(text, k = 4L, hash = NULL)
}
\arguments{
\item{text}{Input vector of text.}

\item{k}{Size of the hash, given as 2^(4*k).}

\item{hash}{Either NULL or cached hash data frame.}
}
\value{
A list containing the data frame of hashed values and the
updated hash table.
}
\description{
Construct a data matrix using hash function.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
