% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch10-implementation.R
\name{casl_fhash_matrix}
\alias{casl_fhash_matrix}
\title{Construct a data matrix using hash function.}
\usage{
casl_fhash_matrix(df, k = 4L)
}
\arguments{
\item{df}{The output of casl_fhash.}

\item{k}{Size of the hash, given as 2^(4*k).}
}
\value{
Sparse matrix representation of the data matrix.
}
\description{
Construct a data matrix using hash function.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
