% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch07-glmnet.R
\name{casl_glenet}
\alias{casl_glenet}
\title{Compute generalized linear elastic net with coordinate descent.}
\usage{
casl_glenet(X, y, lambda, alpha = 1, family = stats::binomial(),
  maxit = 10000L, tol = 1e-05)
}
\arguments{
\item{X}{A numeric data matrix.}

\item{y}{Response vector.}

\item{lambda}{The penalty term.}

\item{alpha}{Value from 0 and 1; balance between l1/l2 penalty.}

\item{family}{Instance of an R `family` object.}

\item{maxit}{Integer maximum number of iterations.}

\item{tol}{Numeric tolerance parameter.}
}
\value{
Regression vector beta of length ncol(X).
}
\description{
Compute generalized linear elastic net with coordinate descent.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
