% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch05-glm.R
\name{casl_glm_nr}
\alias{casl_glm_nr}
\title{Solve generalized linear models with Newton-Ralphson method.}
\usage{
casl_glm_nr(X, y, mu_fun, var_fun, maxit = 25, tol = 1e-10)
}
\arguments{
\item{X}{A numeric data matrix.}

\item{y}{Response vector.}

\item{mu_fun}{Function from eta to the expected value.}

\item{var_fun}{Function from mean to variance.}

\item{maxit}{Integer maximum number of iterations.}

\item{tol}{Numeric tolerance parameter.}
}
\value{
Regression vector beta of length ncol(X).
}
\description{
Solve generalized linear models with Newton-Ralphson method.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
