% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch05-glm.R
\name{casl_glm_nr_logistic}
\alias{casl_glm_nr_logistic}
\title{Logistic regression using the Newton-Ralphson method.}
\usage{
casl_glm_nr_logistic(X, y, maxit = 25L, tol = 1e-10)
}
\arguments{
\item{X}{A numeric data matrix.}

\item{y}{Response vector.}

\item{maxit}{Integer maximum number of iterations.}

\item{tol}{Numeric tolerance parameter.}
}
\value{
Regression vector beta of length ncol(X).
}
\description{
Logistic regression using the Newton-Ralphson method.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
