% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch10-implementation.R
\name{casl_hash_matrix}
\alias{casl_hash_matrix}
\title{Map multiple string objects to a numeric hash values.}
\usage{
casl_hash_matrix(words, k = 4L)
}
\arguments{
\item{words}{A vector of string objects to hash.}

\item{k}{Size of the hash, given as 2^(4*k).}
}
\value{
A data frame giving hash values and signs.
}
\description{
Map multiple string objects to a numeric hash values.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
