% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch10-implementation.R
\name{casl_hash_word2index}
\alias{casl_hash_word2index}
\title{Map a string object to a numeric hash value.}
\usage{
casl_hash_word2index(word, k = 4L)
}
\arguments{
\item{word}{A string object to hash.}

\item{k}{Size of the hash, given as 2^(4*k).}
}
\value{
A numeric hash value from 0 to 2^(4*k) - 1.
}
\description{
Map a string object to a numeric hash value.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
