% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch10-implementation.R
\name{casl_hash_word2sign}
\alias{casl_hash_word2sign}
\title{Map a string object to a sign function.}
\usage{
casl_hash_word2sign(word)
}
\arguments{
\item{word}{A string object to hash.}
}
\value{
A numeric value of either +1 or -1.
}
\description{
Map a string object to a sign function.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
