% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch09-dim.R
\name{casl_kernel_pca}
\alias{casl_kernel_pca}
\title{Compute kernel version of PCA matrix.}
\usage{
casl_kernel_pca(X, k = 2L, kfun = casl_util_kernel_poly, ...)
}
\arguments{
\item{X}{A numeric data matrix.}

\item{k}{Integer number of components to return.}

\item{kfun}{The kernel function.}

\item{...}{Other options passed to the kernel function.}
}
\value{
The nrow(X)-by-k kernel PCA matrix.
}
\description{
Compute kernel version of PCA matrix.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
