% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch07-glmnet.R
\name{casl_lenet_check_kkt}
\alias{casl_lenet_check_kkt}
\title{Check current KKT conditions for regression vector.}
\usage{
casl_lenet_check_kkt(X, y, b, lambda)
}
\arguments{
\item{X}{A numeric data matrix.}

\item{y}{Response vector.}

\item{b}{Current value of the regression vector.}

\item{lambda}{The penalty term.}
}
\value{
A logical vector indicating where the KKT conditions
have been violated by the variables that are currently zero.
}
\description{
Check current KKT conditions for regression vector.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
