% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch07-glmnet.R
\name{casl_lenet_screen}
\alias{casl_lenet_screen}
\title{Apply coordinate descent screening rules.}
\usage{
casl_lenet_screen(X, y, lambda, b = matrix(0, nrow = ncol(X), ncol =
  length(lambda)), maxit = 10000L)
}
\arguments{
\item{X}{A numeric data matrix.}

\item{y}{Response vector.}

\item{lambda}{The penalty term.}

\item{b}{A matrix of warm start coefficients for the algorithm.}

\item{maxit}{Integer maximum number of iterations.}
}
\value{
Named list of parameters for use in the lenet algorithm.
}
\description{
Apply coordinate descent screening rules.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
