% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch07-glmnet.R
\name{casl_lenet_update_beta}
\alias{casl_lenet_update_beta}
\title{Update beta vector using coordinate descent.}
\usage{
casl_lenet_update_beta(X, y, lambda, alpha, b, W)
}
\arguments{
\item{X}{A numeric data matrix.}

\item{y}{Response vector.}

\item{lambda}{The penalty term.}

\item{alpha}{Value from 0 and 1; balance between l1/l2 penalty.}

\item{b}{A vector of warm start coefficients for the algorithm.}

\item{W}{A vector of sample weights.}
}
\value{
A matrix of regression vectors with ncol(X) columns
and length(lambda_vals) rows.
}
\description{
Update beta vector using coordinate descent.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
