% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch07-glmnet.R
\name{casl_lenet_update_beta_kkt}
\alias{casl_lenet_update_beta_kkt}
\title{Update beta vector using KKT conditions.}
\usage{
casl_lenet_update_beta_kkt(X, y, b, lambda, active_set, maxit = 10000L)
}
\arguments{
\item{X}{A numeric data matrix.}

\item{y}{Response vector.}

\item{b}{Current value of the regression vector.}

\item{lambda}{The penalty term.}

\item{active_set}{Logical index of the active set of variables.}

\item{maxit}{Integer maximum number of iterations.}
}
\value{
A list indicating the new regression vector and active set.
}
\description{
Update beta vector using KKT conditions.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
