% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch03-ridge-pca.R
\name{casl_lm_pca}
\alias{casl_lm_pca}
\title{Compute PCA regression vector.}
\usage{
casl_lm_pca(X, y, k)
}
\arguments{
\item{X}{A numeric data matrix.}

\item{y}{Response vector.}

\item{k}{Number of components to use in the fit.}
}
\value{
Regression vector beta of length ncol(X).
}
\description{
Compute PCA regression vector.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
