% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch03-ridge-pca.R
\name{casl_lm_ridge}
\alias{casl_lm_ridge}
\title{Compute ridge regression vector.}
\usage{
casl_lm_ridge(X, y, lambda_vals)
}
\arguments{
\item{X}{A numeric data matrix.}

\item{y}{Response vector.}

\item{lambda_vals}{A sequence of penalty terms.}
}
\value{
A matrix of regression vectors with ncol(X) columns
and length(lambda_vals) rows.
}
\description{
Compute ridge regression vector.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
