% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch04-linear-smoothers.R
\name{casl_nlm1d_kernel}
\alias{casl_nlm1d_kernel}
\title{Apply one-dimensional (Epanechnikov) kernel regression.}
\usage{
casl_nlm1d_kernel(x, y, x_new, h = 1)
}
\arguments{
\item{x}{Numeric vector of the original predictor variables.}

\item{y}{Numeric vector of the responses.}

\item{x_new}{A vector of data values at which to estimate.}

\item{h}{A numeric value giving the bandwidth of the kernel.}
}
\value{
A vector of predictions for each value in x_new.
}
\description{
Apply one-dimensional (Epanechnikov) kernel regression.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
