% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch04-linear-smoothers.R
\name{casl_nlm1d_nat_spline_x}
\alias{casl_nlm1d_nat_spline_x}
\title{Natural cubic spline basis.}
\usage{
casl_nlm1d_nat_spline_x(x, knots)
}
\arguments{
\item{x}{Numeric vector of values.}

\item{knots}{Numeric vector of knot points.}
}
\value{
A matrix with one row for each element of x and
length(knots) columns.
}
\description{
Natural cubic spline basis.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
