% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch04-linear-smoothers.R
\name{casl_nlm1d_poly}
\alias{casl_nlm1d_poly}
\title{Apply one-dimensional polynomial regression.}
\usage{
casl_nlm1d_poly(x, y, n = 1L)
}
\arguments{
\item{x}{Numeric vector of the predictor variables.}

\item{y}{Numeric vector of the responses.}

\item{n}{An integer giving the order to the polynomial.}
}
\value{
A length n numeric vector of coefficients.
}
\description{
Apply one-dimensional polynomial regression.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
