% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch04-linear-smoothers.R
\name{casl_nlm1d_poly_predict}
\alias{casl_nlm1d_poly_predict}
\title{Predict values from one-dimensional polynomial regression.}
\usage{
casl_nlm1d_poly_predict(beta, x, x_new)
}
\arguments{
\item{beta}{A length n numeric vector of coefficients.}

\item{x}{Numeric vector of the original predictor variables.}

\item{x_new}{A vector of data values at which to estimate.}
}
\value{
A vector of predictions with the same length as x_new.
}
\description{
Predict values from one-dimensional polynomial regression.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
