% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch04-linear-smoothers.R
\name{casl_nlm1d_trunc_power_x}
\alias{casl_nlm1d_trunc_power_x}
\title{One-dimensional regression using a truncated power basis.}
\usage{
casl_nlm1d_trunc_power_x(x, knots, order = 1L)
}
\arguments{
\item{x}{Numeric vector of values.}

\item{knots}{Numeric vector of knot points.}

\item{order}{Integer order of the polynomial fit.}
}
\value{
A matrix with one row for each element of x and
(1 + length(knots) + order) columns.
}
\description{
One-dimensional regression using a truncated power basis.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
