% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch06-additive-models.R
\name{casl_nlm_poly}
\alias{casl_nlm_poly}
\title{Nonlinear regression using polynomial regression.}
\usage{
casl_nlm_poly(X, y, X_new, k)
}
\arguments{
\item{X}{A numeric data matrix.}

\item{y}{Response vector.}

\item{X_new}{Numeric data matrix of prediction locations.}

\item{k}{Order of the polynomial.}
}
\value{
A vector of predictions; one for each row in X_new.
}
\description{
Nonlinear regression using polynomial regression.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
