% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch08-neural-networks.R
\name{casl_nn_grad_check}
\alias{casl_nn_grad_check}
\title{Perform a gradient check for the dense NN code.}
\usage{
casl_nn_grad_check(X, y, weights, h = 1e-04)
}
\arguments{
\item{X}{A numeric data matrix.}

\item{y}{A numeric vector of responses.}

\item{weights}{List of weights describing the neural network.}

\item{h}{Positive numeric bandwidth to test.}
}
\value{
The largest difference between the empirical and analytic
gradients of the weights and biases.
}
\description{
Perform a gradient check for the dense NN code.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
