% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch08-neural-networks.R
\name{casl_nnmulti_forward_prop}
\alias{casl_nnmulti_forward_prop}
\title{Apply forward propagation to for a multinomial NN.}
\usage{
casl_nnmulti_forward_prop(x, weights, sigma)
}
\arguments{
\item{x}{A numeric vector representing one row of the input.}

\item{weights}{A list created by casl_nn_make_weights.}

\item{sigma}{The activation function.}
}
\value{
A list containing the new weighted responses (z) and activations (a).
}
\description{
Apply forward propagation to for a multinomial NN.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
