% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch08-neural-networks.R
\name{casl_nnmulti_predict}
\alias{casl_nnmulti_predict}
\title{Predict values from training a multinomial neural network.}
\usage{
casl_nnmulti_predict(weights, X_test)
}
\arguments{
\item{weights}{List of weights describing the neural network.}

\item{X_test}{A numeric data matrix for the predictions.}
}
\value{
A matrix of predicted values.
}
\description{
Predict values from training a multinomial neural network.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
