% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch02-linear-models.R
\name{casl_ols_orth_proj}
\alias{casl_ols_orth_proj}
\title{Compute OLS estimate using the orthogonal projection.}
\usage{
casl_ols_orth_proj(X, y)
}
\arguments{
\item{X}{A numeric data matrix.}

\item{y}{Response vector.}
}
\value{
Regression vector beta of length ncol(X).
}
\description{
Compute OLS estimate using the orthogonal projection.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
