% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch10-implementation.R
\name{casl_sparse_add}
\alias{casl_sparse_add}
\title{Add two sparse matrices stored as triplet lists.}
\usage{
casl_sparse_add(a, b)
}
\arguments{
\item{a}{A list describing a sparse matrix.}

\item{b}{A list describing a sparse matrix.}
}
\value{
A list describing the sum of a and b as a sparse matrix.
}
\description{
Add two sparse matrices stored as triplet lists.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
