% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch10-implementation.R
\name{casl_sparse_lmax}
\alias{casl_sparse_lmax}
\title{Compute starting lambda for lasso regression.}
\usage{
casl_sparse_lmax(X, y)
}
\arguments{
\item{X}{A sparse or dense numeric matrix.}

\item{y}{The numeric response vector.}
}
\value{
The numeric value of lambda max.
}
\description{
Compute starting lambda for lasso regression.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
