% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch09-dim.R
\name{casl_spectral_clust}
\alias{casl_spectral_clust}
\title{Perform spectral clustering using a similarity matrix.}
\usage{
casl_spectral_clust(S, k = 1)
}
\arguments{
\item{S}{A symmetric similarity matrix.}

\item{k}{Number of components to return.}
}
\value{
A matrix with nrow(S) rows and k columns.
}
\description{
Perform spectral clustering using a similarity matrix.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
