% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch09-dim.R
\name{casl_tsne}
\alias{casl_tsne}
\title{Compute t-SNE variance values.}
\usage{
casl_tsne(X, perplexity = 30, k = 2L, iter = 1000L, rho = 100)
}
\arguments{
\item{X}{A numeric data matrix.}

\item{perplexity}{Desired perplexity score for all variables.}

\item{k}{Dimensionality of the output.}

\item{iter}{Number of iterations to perform.}

\item{rho}{A positive numeric learning rate.}
}
\value{
An nrow(X) by k matrix of t-SNE embeddings.
}
\description{
Compute t-SNE variance values.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
