% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch09-dim.R
\name{casl_tsne_p}
\alias{casl_tsne_p}
\title{Compute t-SNE probability scores.}
\usage{
casl_tsne_p(X, perplexity = 15)
}
\arguments{
\item{X}{A numeric data matrix.}

\item{perplexity}{Desired perplexity score for all variables.}
}
\value{
A matrix of probability densities.
}
\description{
Compute t-SNE probability scores.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
