% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch09-dim.R
\name{casl_tsne_sigma}
\alias{casl_tsne_sigma}
\title{Compute t-SNE variance values.}
\usage{
casl_tsne_sigma(X, perplexity)
}
\arguments{
\item{X}{A numeric data matrix.}

\item{perplexity}{Desired perplexity score for all variables.}
}
\value{
A vector of estimated variances for each variable.
}
\description{
Compute t-SNE variance values.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
