% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch08-neural-networks.R
\name{casl_util_conv}
\alias{casl_util_conv}
\title{Apply the convolution operator.}
\usage{
casl_util_conv(x, w)
}
\arguments{
\item{x}{The input image as a matrix.}

\item{w}{Matrix of the kernel weight.}
}
\value{
Vector of the output convolution.
}
\description{
Apply the convolution operator.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
