% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch04-linear-smoothers.R
\name{casl_util_kernel_epan}
\alias{casl_util_kernel_epan}
\title{Evaluate the Epanechnikov kernel function.}
\usage{
casl_util_kernel_epan(x, h = 1)
}
\arguments{
\item{x}{Numeric vector of points to evaluate the function at.}

\item{h}{A numeric value giving the bandwidth of the kernel.}
}
\value{
A vector of values with the same length as x.
}
\description{
Evaluate the Epanechnikov kernel function.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
