% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch09-dim.R
\name{casl_util_kernel_norm}
\alias{casl_util_kernel_norm}
\title{Calculate normalized kernel matrix.}
\usage{
casl_util_kernel_norm(M)
}
\arguments{
\item{M}{A raw kernel matrix.}
}
\value{
A normalized version of the kernel matrix.
}
\description{
Calculate normalized kernel matrix.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
