% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch09-dim.R
\name{casl_util_kernel_poly}
\alias{casl_util_kernel_poly}
\title{Compute square kernel matrix for polynomial kernel.}
\usage{
casl_util_kernel_poly(X, d = 2L, c = 0)
}
\arguments{
\item{X}{A numeric data matrix.}

\item{d}{Integer degree of the polynomial.}

\item{c}{Numeric constant to modify the kernel.}
}
\value{
A square kernel matrix with the nrow(X) rows and columns.
}
\description{
Compute square kernel matrix for polynomial kernel.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
