% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch09-dim.R
\name{casl_util_kernel_radial}
\alias{casl_util_kernel_radial}
\title{Compute square kernel matrix for a radial kernel.}
\usage{
casl_util_kernel_radial(X, gamma = 1)
}
\arguments{
\item{X}{A numeric data matrix.}

\item{gamma}{Positive tuning parameter to set kernel shape.}
}
\value{
A square kernel matrix with the nrow(X) rows and columns.
}
\description{
Compute square kernel matrix for a radial kernel.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
