% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch09-dim.R
\name{casl_util_knn_sim}
\alias{casl_util_knn_sim}
\title{Compute symmetric k-nearest neighbors similarity matrix.}
\usage{
casl_util_knn_sim(X, k = 10L)
}
\arguments{
\item{X}{A numeric data matrix.}

\item{k}{Number of nearest neighbors to consider.}
}
\value{
A square similarity matrix with nrow(X) rows & columns.
}
\description{
Compute symmetric k-nearest neighbors similarity matrix.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
