% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch08-neural-networks.R
\name{casl_util_mse_p}
\alias{casl_util_mse_p}
\title{Derivative of the mean squared error (MSE) function.}
\usage{
casl_util_mse_p(y, a)
}
\arguments{
\item{y}{A numeric vector of responses.}

\item{a}{A numeric vector of predicted responses.}
}
\value{
Returned current derivative the MSE function.
}
\description{
Derivative of the mean squared error (MSE) function.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
