% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch08-neural-networks.R
\name{casl_util_ReLU}
\alias{casl_util_ReLU}
\title{Apply a rectified linear unit (ReLU) to a vector/matrix.}
\usage{
casl_util_ReLU(v)
}
\arguments{
\item{v}{A numeric vector or matrix.}
}
\value{
The original input with negative values truncated to zero.
}
\description{
Apply a rectified linear unit (ReLU) to a vector/matrix.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
