% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch08-neural-networks.R
\name{casl_util_ReLU_p}
\alias{casl_util_ReLU_p}
\title{Apply derivative of the rectified linear unit (ReLU).}
\usage{
casl_util_ReLU_p(v)
}
\arguments{
\item{v}{A numeric vector or matrix.}
}
\value{
Sets positive values to 1 and negative values to zero.
}
\description{
Apply derivative of the rectified linear unit (ReLU).
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
