% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch03-ridge-pca.R
\name{casl_util_rmse}
\alias{casl_util_rmse}
\title{Root mean squared error of a prediction.}
\usage{
casl_util_rmse(y, y_hat, by)
}
\arguments{
\item{y}{A vector of responses to predict.}

\item{y_hat}{A vector of predicted values.}

\item{by}{Optional categories to stratify results.}
}
\value{
A table of the root mean squared error for each
group in "by".
}
\description{
Root mean squared error of a prediction.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
