% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch07-glmnet.R
\name{casl_util_soft_thresh}
\alias{casl_util_soft_thresh}
\title{Soft threshold function.}
\usage{
casl_util_soft_thresh(a, b)
}
\arguments{
\item{a}{Numeric vector of values to threshold.}

\item{b}{The soft thresholded value.}
}
\value{
Numeric vector of the soft-thresholded values of a.
}
\description{
Soft threshold function.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
