% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch08-neural-networks.R
\name{casl_util_softmax}
\alias{casl_util_softmax}
\title{Apply the softmax function to a vector.}
\usage{
casl_util_softmax(z)
}
\arguments{
\item{z}{A numeric vector of inputs.}
}
\value{
Output after applying the softmax function.
}
\description{
Apply the softmax function to a vector.
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
