% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch01-introduction.R
\name{casl_utils_best_split}
\alias{casl_utils_best_split}
\title{Run the annotation pipeline on a set of documents}
\usage{
casl_utils_best_split(x, y)
}
\arguments{
\item{x}{Numeric vector with which to classify the data.}

\item{y}{Vector of responses coded as 0s and 1s.}
}
\value{
The best split value(s).
}
\description{
Run the annotation pipeline on a set of documents
}
\references{
Taylor Arnold, Michael Kane, and Bryan Lewis.
\emph{A Computational Approach to Statistical Learning}.
Chapman & Hall/CRC Texts in Statistical Science, 2019.
}
\author{
Taylor Arnold, Michael Kane, Bryan Lewis.
}
