% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{data}
\name{reviews}
\alias{reviews}
\title{A Sample of Amazon Book Reviews from Y2k}
\description{
A sample of reviews of books sold by Amazon from the year 2000 includes a
review identifier, asin, reviewer name, the helpfulness of the review,
the review text, the overall score of the product, a summary, the
unix formatted date of the review, and the date of the review.

The data set is used to illustrate the generalized, penalized 
linear regression in Computational Approaches to Statistical Learning.
}
\references{
\url{http://snap.stanford.edu/data/amazon/productGraph/categoryFiles/reviews_Books_5.json.gz}

\url{http://jmcauley.ucsd.edu/data/amazon/}
}
